Visual Studio Git add-in is excluding files
Apparently at random, the VS add in for Git is choosing not to include
perfectly reasonable files in the repository. This commit is simply to
include a number of those files, and will pick up some new changes as
collateral damage.  Better to get the commit done now rather than mix
the correction with a lot of real changes.