Added framework classes for console applications
Added a framework class for command driven apps, and another for the
usual type of console application. Came up with a way to get the
framework class to run with minimal overhead. Attempted to discourage
piling all of the code into a Program class by providing ways to place
the functionality of the program into command classes or dedicated
handler classes. The Program class can still support all of the commands
directly, but it is no more work to divide the code up into dedicated
classes.