Improved input recognition.
In particular, resolution of the modified F3 key vs CPR (cursor position report) which can look identical (F3 resembles a CPR where the cursor is on row 1). In order to receive modified F3 (Ctrl-F3, Shift-F3 etc) a setting needs to be supplied to indicate that row 1 CPRs should be interpreted as the F3 key.